incorporate multipart context impl into existing context impl
